# utils/swaggerDoc.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Menu api",
    "description": "This api is about a food's menu, you can create a user, then see the food that we have and if you want buy some food create a bill.",
    "contact": {
      "name": "Github",
      "url": "https://github.com/JuanCamilo08/menu-api"
    },
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
...
  "paths": {
...
    "/api/ingredient": {
      "get": {
        "tags": ["Ingredient"],
        "description": "Get all the ingredients.",
        "parameters": [
          {
            "name": "x-auth-token",
            "in": "header",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "string"
            },
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc0FkbWluIjp0cnVlLCJfaWQiOiI1ZTYwN2VkNTMxYWYzNjJmMTU1OTM0MGYiLCJpYXQiOjE1ODM0MTU1NzV9.RON1F9inuwkExyu2zdw9-3-_4Djba5g4MefqcQ9KiIY"
          }
        ],
        "responses": {
          "200": {
            "description": "return an array of ingredient's objects.",
            "content": {
              "application/json; charset=utf-8": {
                "schema": {
                  "type": "string"
                },
                "examples": {
                  "1": {
                    "value": "[\n  {\n    \"_id\": \"5e591f0216d7745d4578841f\",\n     \"name\": \"name of ingredient\",\n    \"__v\": 0\n  },\n  {\n    \"_id\": \"5e591f0b16d7745d45788420\",\n    \"name\": \"name of ingredient\",\n    \"__v\": 0\n  }\n]"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Ingredient"],
        "description": "Create a new ingredient.",
        "parameters": [
          {
            "name": "x-auth-token",
            "in": "header",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "string"
            },
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc0FkbWluIjp0cnVlLCJfaWQiOiI1ZTYwN2VkNTMxYWYzNjJmMTU1OTM0MGYiLCJpYXQiOjE1ODM0MTU1NzV9.RON1F9inuwkExyu2zdw9-3-_4Djba5g4MefqcQ9KiIY"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/body_2"
              },
              "examples": {
                "1": {
                  "value": "{\n    \"name\": \"new ingredient\"\n}"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "It returns the created ingredient.",
            "content": {
              "application/json; charset=utf-8": {
                "schema": {
                  "type": "string"
                },
                "examples": {
                  "1": {
                    "value": "{\n  \"_id\": \"5e591f0216d7745d4578841f\",\n  \"name\": \"new ingredient\",\n  \"__v\": 0\n}"
                  }
                }
              }
            }
          }
        },
        "servers": [
          {
            "url": "http://localhost:3000"
          }
        ]
      },
      "servers": [
        {
          "url": "http://localhost:3000"
        }
      ]
    },
    "/api/ingredient/{ingredientId}": {
      "put": {
        "tags": ["Ingredient"],
        "description": "Update an already existing ingredient.",
        "parameters": [
          {
            "name": "ingredientId",
            "in": "path",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "string"
            },
            "example": "5e6101c4eb52945744cdc9f1"
          },
          {
            "name": "x-auth-token",
            "in": "header",
            "required": false,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "string"
            },
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc0FkbWluIjp0cnVlLCJfaWQiOiI1ZTYwN2VkNTMxYWYzNjJmMTU1OTM0MGYiLCJpYXQiOjE1ODM0MTU1NzV9.RON1F9inuwkExyu2zdw9-3-_4Djba5g4MefqcQ9KiIY"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/body_3"
              },
              "examples": {
                "1": {
                  "value": "{\n    \"name\": \"ingredient's name\"\n}"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Update the ingredient and return the new one.",
            "content": {
              "application/json; charset=utf-8": {
                "schema": {
                  "type": "string"
                },
                "examples": {
                  "1": {
                    "value": "{\n  \"_id\": \"5e591f0216d7745d4578841f\",\n  \"name\": \"ingredient's name\",\n  \"__v\": 0\n}"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not ingredient found.",
            "content": {
              "text/html; charset=utf-8": {
                "schema": {
                  "type": "string"
                },
                "examples": {
                  "1": {
                    "value": "The ingredient with the given ID was not found."
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Ingredient"],
        "description": "Delete an ingredient by his Id.",
        "parameters": [
          {
            "name": "ingredientId",
            "in": "path",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "string"
            },
            "example": "5e6101c4eb52945744cdc9f1"
          },
          {
            "name": "x-auth-token",
            "in": "header",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "string"
            },
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc0FkbWluIjp0cnVlLCJfaWQiOiI1ZTYwN2VkNTMxYWYzNjJmMTU1OTM0MGYiLCJpYXQiOjE1ODM0MTU1NzV9.RON1F9inuwkExyu2zdw9-3-_4Djba5g4MefqcQ9KiIY"
          }
        ],
        "responses": {
          "200": {
            "description": "Return the object of the ingredient deleted.",
            "content": {
              "application/json; charset=utf-8": {
                "schema": {
                  "type": "string"
                },
                "examples": {
                  "1": {
                    "value": "{\n  \"_id\": \"5e591f0216d7745d4578841f\",\n  \"name\": \"ingredient's name\",\n  \"__v\": 0\n}"
                  }
                }
              }
            }
          },
          "404": {
            "description": "The ingredient does not exist.",
            "content": {
              "text/html; charset=utf-8": {
                "schema": {
                  "type": "string"
                },
                "examples": {
                  "1": {
                    "value": "The ingredient with the given ID was not found."
                  }
                }
              }
            }
          }
        },
        "servers": [
          {
            "url": "http://localhost:3000"
          }
        ]
      },
      "servers": [
        {
          "url": "http://localhost:3000"
        }
      ]
    },
...
  },
...
}